+++ persistence/src/main/java/org/devgateway/toolkit/persistence/dao/form/AwardNotification.java
@@ -12,6 +12,7 @@ import javax.persistence.Entity;
 import javax.persistence.Index;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
@@ -28,7 +29,7 @@ public class AwardNotification extends AbstractPurchaseReqMakueniEntity {
     private Date awardDate;
 
     @ExcelExport(useTranslation = true)
-    private Double tenderValue;
+    private BigDecimal tenderValue;
 
     @ExcelExport(justExport = true, useTranslation = true)
     @Cache(usage = CacheConcurrencyStrategy.READ_WRITE)
@@ -38,11 +39,11 @@ public class AwardNotification extends AbstractPurchaseReqMakueniEntity {
     @ExcelExport(useTranslation = true)
     private Integer acknowledgementDays;
 
-    public Double getTenderValue() {
+    public BigDecimal getTenderValue() {
         return tenderValue;
     }
 
-    public void setTenderValue(final Double tenderValue) {
+    public void setTenderValue(final BigDecimal tenderValue) {
         this.tenderValue = tenderValue;
     }
